refactor for all files
still a lil buggy
