+++ source/local/imports.js
@@ -1,0 +1,16 @@
+// CLASSES
+export * from "./classes/TypeA.js";
+export * from "./classes/TypeF.js";
+export * from "./classes/AudioObj.js";
+
+// FUNCTIONS
+export * as notes_utils from "./scripts/notes.js";
+export * as data_utils from "./scripts/data.js";
+export * as audio_utils from "./scripts/audio.js";
+
+// UTILS
+export * as utils from "./scripts/utils.js";
+
+// CONSTANTS
+export const dict_typeFs = {};
+export const DEBUG_FLAG = true;
