+++ source/local/scripts/imports.js
@@ -1,0 +1,12 @@
+// CLASSES
+export * from "../classes/TypeA.js";
+export * from "../classes/TypeF.js";
+export * from "../classes/AudioObj.js";
+
+// FUNCTIONS
+export * as notes from "./notes.js"
+export * as data from "./data.js"
+export * as audio from "./audio.js"
+
+// UTILS
+export * as utils from "./utils.js"
